requirements.txt
Flask==3.0.3
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
